// packages/office-ui-fabric-react/src/packages/react-pickers/components/SwatchColorPicker/ColorPickerGridCell.types.ts
export interface IColorCellProps {
  /**
   * Arbitrary unique string associated with this option
   */
  id: string;

  /**
   * The label for this item.
   * Visible text if this item is a header,
   * tooltip if is this item is normal
   */
  label?: string;

  /**
   * The CSS-compatible string to describe the color
   */
  color?: string;

  /**
   * Index for this option
   */
  index?: number;
}